Setup initial mdx layout and packages
In order to start making some posts for training we need a basic layout
to work with. I choose 'full-bleed' layout since these are similar to
blog posts.

The fonts are close to what would be seen in an live training session so
they mimic handwritting, but keep the eligibility of non-handwritten
fonts.

More updates will be done in the future to setup header, footer, etc.