Add Service Design Session and Introduction
When read on the page without the header it looked weird so it was redesigned to
include a header that at least has some framing.

All elements were defined only for the session probably it will be worth
exporting them and importing them for other sessions.

The introduction is nothing fancy, dynamic behaviour will be added
later.